condensed footer into one component,readme edits, some style edits.
