+++ src/routes/Home.tsx
@@ -71,12 +71,6 @@ const StyledDivider = Styled(Divider)({
   marginTop: "5em !important"
 });
 
-const StyledFooter = Styled(Segment)({
-  padding: "5em 0em !important",
-  marginTop: "70px !important",
-  background: "#08A6C9 !important"
-});
-
 function Home() {
   const { authState } = useOktaAuth();
 
@@ -221,55 +215,12 @@ function Home() {
     );
   };
 
-  const HomePageFooter = () => {
-    return (
-      <div>
-        <StyledFooter inverted vertical>
-          <Container>
-            <Grid divided inverted stackable>
-              <Grid.Row>
-                <Grid.Column width={3}>
-                  <Header inverted as="h4" content="About" />
-                  <List link inverted>
-                    <List.Item as="a">Sitemap</List.Item>
-                    <List.Item as="a">Contact Us</List.Item>
-                    <List.Item as="a">Religious Ceremonies</List.Item>
-                    <List.Item as="a">Gazebo Plans</List.Item>
-                  </List>
-                </Grid.Column>
-                <Grid.Column width={3}>
-                  <Header inverted as="h4" content="Services" />
-                  <List link inverted>
-                    <List.Item as="a">Banana Pre-Order</List.Item>
-                    <List.Item as="a">DNA FAQ</List.Item>
-                    <List.Item as="a">How To Access</List.Item>
-                    <List.Item as="a">Favorite X-Men</List.Item>
-                  </List>
-                </Grid.Column>
-                <Grid.Column width={7}>
-                  <Header as="h4" inverted>
-                    Footer Header
-                  </Header>
-                  <p>
-                    Extra space for a call to action inside the footer that
-                    could help re-engage users.
-                  </p>
-                </Grid.Column>
-              </Grid.Row>
-            </Grid>
-          </Container>
-        </StyledFooter>
-      </div>
-    );
-  };
-
   return (
     <div>
       {!authState.isAuthenticated ? (
         <div>
           <HomePageHeader />
           <HomePageBody />
-          <HomePageFooter />
         </div>
       ) : (
         <Redirect to="/dashboard" />
