+++ src/routes/Protected/dashboard.tsx
@@ -25,12 +25,6 @@ const StyledBackGround = Styled.div`
   background: #F3F8F9;
 `;
 
-const StyledSegment = Styled(Segment)({
-  padding: "5em 0em !important",
-  background: "#08A6C9 !important",
-  marginTop: "10em !important"
-});
-
 const Dashboard = () => {
   const [jobs, setJobs] = useState([]);
   const [loading, setLoading] = useState(true);
@@ -78,7 +72,7 @@ const Dashboard = () => {
       );
     } else {
       return (
-        <div>
+        <div style={{ minHeight: "50vh" }}>
           <Grid stackable container columns="equal">
             <Grid.Row stretched>
               {jobs.length > 0 ? (
@@ -104,42 +98,6 @@ const Dashboard = () => {
       }`}</StyledHeader>
 
       <Loading />
-
-      <StyledSegment inverted vertical>
-        <Container>
-          <Grid divided inverted stackable>
-            <Grid.Row>
-              <Grid.Column width={3}>
-                <Header inverted as="h4" content="About" />
-                <List link inverted>
-                  <List.Item as="a">Sitemap</List.Item>
-                  <List.Item as="a">Contact Us</List.Item>
-                  <List.Item as="a">Religious Ceremonies</List.Item>
-                  <List.Item as="a">Gazebo Plans</List.Item>
-                </List>
-              </Grid.Column>
-              <Grid.Column width={3}>
-                <Header inverted as="h4" content="Services" />
-                <List link inverted>
-                  <List.Item as="a">Banana Pre-Order</List.Item>
-                  <List.Item as="a">DNA FAQ</List.Item>
-                  <List.Item as="a">How To Access</List.Item>
-                  <List.Item as="a">Favorite X-Men</List.Item>
-                </List>
-              </Grid.Column>
-              <Grid.Column width={7}>
-                <Header as="h4" inverted>
-                  Footer Header
-                </Header>
-                <p>
-                  Extra space for a call to action inside the footer that could
-                  help re-engage users.
-                </p>
-              </Grid.Column>
-            </Grid.Row>
-          </Grid>
-        </Container>
-      </StyledSegment>
     </StyledBackGround>
   );
 };
