+++ src/routes/footer.tsx
@@ -1,26 +1,51 @@
 import React from "react";
 import Styled from "styled-components";
-
-const FooterDiv = Styled.div`
-width:100% !important;
-background:#08A6C9;
-color:white;
-height:40px;
-z-index:2;
-bottom:0 !important;
-position: relative;
-text-align: center;
-margin-top: 500px;
-
-`;
+import { Segment, Container, Grid, Header, List } from "semantic-ui-react";
+const StyledFooter = Styled(Segment)({
+  padding: "5em 0em !important",
+  marginTop: "70px !important",
+  background: "#08A6C9 !important"
+});
 
 function Footer() {
   return (
-    <>
-      <FooterDiv>
-        <h1>Save this Job</h1>
-      </FooterDiv>
-    </>
+    <div>
+      <StyledFooter inverted vertical>
+        <Container>
+          <Grid divided inverted stackable>
+            <Grid.Row>
+              <Grid.Column width={3}>
+                <Header inverted as="h4" content="About" />
+                <List link inverted>
+                  <List.Item as="a">Sitemap</List.Item>
+                  <List.Item as="a">Contact Us</List.Item>
+                  <List.Item as="a">Religious Ceremonies</List.Item>
+                  <List.Item as="a">Gazebo Plans</List.Item>
+                </List>
+              </Grid.Column>
+              <Grid.Column width={3}>
+                <Header inverted as="h4" content="Services" />
+                <List link inverted>
+                  <List.Item as="a">Banana Pre-Order</List.Item>
+                  <List.Item as="a">DNA FAQ</List.Item>
+                  <List.Item as="a">How To Access</List.Item>
+                  <List.Item as="a">Favorite X-Men</List.Item>
+                </List>
+              </Grid.Column>
+              <Grid.Column width={7}>
+                <Header as="h4" inverted>
+                  Footer Header
+                </Header>
+                <p>
+                  Extra space for a call to action inside the footer that could
+                  help re-engage users.
+                </p>
+              </Grid.Column>
+            </Grid.Row>
+          </Grid>
+        </Container>
+      </StyledFooter>
+    </div>
   );
 }
 export default Footer;
