+++ src/semantic-ui/site/globals/site.overrides
@@ -1,12 +1,4 @@
 /*******************************
          Site Overrides
 *******************************/
-.ui[class*="vertically divided"].grid > .row:before {
-    position: absolute;
-    content: "";
-    top: 0em;
-    left: 0px;
-    width: calc(100% - 2rem);
-    height: 1px;
-    margin: 0% 1rem;
-}
+
