+++ src/semantic-ui/theme.config
@@ -22,7 +22,7 @@
 @reset       : 'default';
 
 /* Elements */
-@button      : 'default';
+@button      : 'twitter';
 @container   : 'default';
 @divider     : 'default';
 @flag        : 'default';
